fix(MP-74) - Rename folder to reflect component name.
